refactor: Extract Book
Extract book as stateless functional component; this is a div not an list item. The property parameter is a book details object.

Add a static array, "books", to Bookshelf. This is to support the static
version "as-is". Bookshelf should be renamed; perhaps Bookrack or
ListBooksContent. The former has better metaphorical value, while the
latter conforms more strongly with the CSS naming convention.